Add a layer of cache for the number of overflowing tracks
This reduces the number of API requests by a large amount as it no longer checks the number of overflowing tracks after each addition. A balance is maintained locally for each playlist. This balance is checked against the API once every 12 hours to make sure any manual change is detected and corrected.