Update radio.net API
radio.net has updated their API. The new API seems to be a lot worse – there's no way to specify how many tracks in the history request. There's also no way to request history for multiple stations at once. The only good thing is that we no longer need to provide an API key to use it (for now).